net: lib: nrf_cloud_fota: enable SMP FOTA type
Enable SMP FOTA type for MQTT-based FOTA.

IRIS-9480
IRIS-9482

+++ subsys/net/lib/nrf_cloud/src/nrf_cloud_fota.c
@@ -67,9 +67,14 @@ static int save_validate_status(const char *const job_id,
 			   const enum nrf_cloud_fota_validate_status status);
 static int publish_validated_job_status(void);
 static void reset_topics(void);
+static void install_smp_update(struct nrf_cloud_fota_job *fota);
 
 static struct mqtt_client *client_mqtt;
 static nrf_cloud_fota_callback_t event_cb;
+
+#if defined(CONFIG_NRF_CLOUD_FOTA_SMP)
+static dfu_target_reset_cb_t smp_reset_cb;
+#endif
 
 static struct mqtt_topic topic_updt = { .qos = MQTT_QOS_1_AT_LEAST_ONCE };
 static struct mqtt_topic topic_req = { .qos = MQTT_QOS_1_AT_LEAST_ONCE };
@@ -142,6 +147,10 @@ static void send_fota_done_event_if_done(void)
 	 */
 	if ((current_fota.status == NRF_CLOUD_FOTA_IN_PROGRESS) &&
 	    (saved_job.validate == NRF_CLOUD_FOTA_VALIDATE_PENDING)) {
+		/* SMP updates can be installed immediately */
+		if (current_fota.info.type == NRF_CLOUD_FOTA_SMP) {
+			install_smp_update(&current_fota);
+		}
 		send_event(NRF_CLOUD_FOTA_EVT_DONE, &current_fota);
 	}
 }
@@ -157,9 +166,10 @@ static int pending_fota_job_validate(void)
 	}
 
 	/* Do not enforce a reboot for modem updates since they can also be handled by
-	 * reinitializing the modem lib
+	 * reinitializing the modem lib. SMP and custom updates also do not need to reboot.
 	 */
-	if (!nrf_cloud_fota_is_type_modem(saved_job.type)) {
+	if (!nrf_cloud_fota_is_type_modem(saved_job.type) &&
+	    (saved_job.type != NRF_CLOUD_FOTA_SMP) && (saved_job.type != NRF_CLOUD_FOTA_CUSTOM)) {
 		reboot_on_init = reboot;
 	}
 
@@ -211,7 +221,7 @@ static void fota_reboot(void)
 	sys_reboot(SYS_REBOOT_COLD);
 }
 
-int nrf_cloud_fota_init(nrf_cloud_fota_callback_t cb)
+int nrf_cloud_fota_init(struct nrf_cloud_fota_init_param const *const init)
 {
 	int ret;
 
@@ -219,12 +229,16 @@ int nrf_cloud_fota_init(nrf_cloud_fota_callback_t cb)
 		fota_reboot();
 	}
 
-	if (cb == NULL) {
+	if (init == NULL) {
 		LOG_ERR("Invalid parameter");
 		return -EINVAL;
 	}
 
-	event_cb = cb;
+	event_cb = init->evt_cb;
+
+#if defined(CONFIG_NRF_CLOUD_FOTA_SMP)
+	smp_reset_cb = init->smp_reset_cb;
+#endif
 
 	if (initialized) {
 		return 0;
@@ -239,6 +253,14 @@ int nrf_cloud_fota_init(nrf_cloud_fota_callback_t cb)
 			LOG_ERR("fota_download_init error: %d", ret);
 			return ret;
 		}
+#if defined(CONFIG_NRF_CLOUD_FOTA_SMP)
+		ret = mcumgr_smp_client_init(smp_reset_cb);
+		if (ret != 0) {
+			LOG_ERR("mcumgr_smp_client_init error: %d", ret);
+			return ret;
+		}
+		(void)nrf_cloud_fota_smp_version_read();
+#endif
 		fota_dl_initialized = true;
 	}
 
@@ -621,6 +643,32 @@ static int save_validate_status(const char *const job_id,
 	return ret;
 }
 
+static void install_smp_update(struct nrf_cloud_fota_job *fota)
+{
+	enum nrf_cloud_fota_validate_status validate = NRF_CLOUD_FOTA_VALIDATE_UNKNOWN;
+	int ret = nrf_cloud_fota_smp_install();
+
+	if (ret == 0) {
+		validate = NRF_CLOUD_FOTA_VALIDATE_PASS;
+		current_fota.status = NRF_CLOUD_FOTA_SUCCEEDED;
+	} else {
+		validate = NRF_CLOUD_FOTA_VALIDATE_FAIL;
+		current_fota.status = NRF_CLOUD_FOTA_FAILED;
+		current_fota.error = NRF_CLOUD_FOTA_ERROR_APPLY_FAIL;
+	}
+
+	/* TODO: after installation, read the version number.
+	 * This does not work for some reason.
+	 * There seems to be some sort of crash/error in dfu_target_smp_image_list_get()
+	 * when called after image installation.
+	 * For now, just reboot everything and read the version on startup.
+	 */
+
+	save_validate_status(current_fota.info.id,
+			     current_fota.info.type,
+			     validate);
+}
+
 static void http_fota_handler(const struct fota_download_evt *evt)
 {
 	__ASSERT_NO_MSG(evt != NULL);
@@ -648,7 +696,9 @@ static void http_fota_handler(const struct fota_download_evt *evt)
 		save_validate_status(current_fota.info.id,
 				     current_fota.info.type,
 				     NRF_CLOUD_FOTA_VALIDATE_PENDING);
+
 		ret = publish_job_status(&current_fota);
+
 		break;
 
 	case FOTA_DOWNLOAD_EVT_ERASE_PENDING:
@@ -816,6 +866,9 @@ static int start_job(struct nrf_cloud_fota_job *const job, const bool send_evt)
 			ret = -EFTYPE;
 		}
 		break;
+	case NRF_CLOUD_FOTA_SMP:
+		img_type = DFU_TARGET_IMAGE_TYPE_SMP;
+		break;
 	default:
 		LOG_ERR("Unhandled FOTA type: %d", job->info.type);
 		ret = -EFTYPE;
@@ -839,7 +892,11 @@ static int start_job(struct nrf_cloud_fota_job *const job, const bool send_evt)
 			.pdn_id = 0,
 			.frag_size_override = CONFIG_NRF_CLOUD_FOTA_DOWNLOAD_FRAGMENT_SIZE,
 		},
-		.fota = { .expected_type = img_type }
+		.fota = {
+			.expected_type = img_type,
+			.img_sz = job->info.file_size,
+			.cb = (void *)http_fota_handler
+		}
 	};
 
 	ret = nrf_cloud_download_start(&dl);
